backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
Django==6.0.3
django-celery-beat==2.9.0
django-environ==0.13.0
django-filter==25.2
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
frozenlist==1.8.0
google-auth==2.49.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
Markdown==3.10.2
msgpack==1.1.2
multidict==6.7.1
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyfcm==2.1.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qrcode==8.2
redis==7.3.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
twilio==9.10.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
yarl==1.23.0
zope.interface==8.2
